Feature/ajout services

+++ src/main/java/com/dsi/projetgestionpfe/repositories/JuryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JuryRepository extends JpaRepository<Jury, Long> {
+public interface JuryRepository extends JpaRepository<Jury, Integer> {
 }
