+++ src/main/java/com/dsi/projetgestionpfe/repositories/RapportRepository.java
@@ -5,6 +5,9 @@ import com.dsi.projetgestionpfe.entities.Rapport;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
+
 @Repository
-public interface RapportRepository extends JpaRepository<Rapport, Long> {
+public interface RapportRepository extends JpaRepository<Rapport, Integer> {
+
 }
